refactor: extract flow init helpers
