+++ packages/core/src/view/viewUtils.js
@@ -3,8 +3,6 @@ import { isObject, isString } from "vega-util";
 import UnitView from "./unitView.js";
 // eslint-disable-next-line no-unused-vars
 import View, { VISIT_SKIP, VISIT_STOP } from "./view.js";
-import { buildDataFlow } from "./flowBuilder.js";
-import { optimizeDataFlow } from "../data/flowOptimizer.js";
 import { isFieldDef, primaryPositionalChannels } from "../encoder/encoder.js";
 import { rollup } from "d3-array";
 import { concatUrl, getDirectory } from "../utils/url.js";
@@ -90,111 +88,6 @@ export function setImplicitScaleNames(root) {
             resolution.name = `${channel}_at_root`;
         }
     }
-}
-
-/**
- * @param {View} root
- * @param {import("../data/dataFlow.js").default<View>} [existingFlow] Add data flow
- *      graphs to an existing DataFlow object.
- */
-export async function initializeData(root, existingFlow) {
-    const flow = buildDataFlow(root, existingFlow);
-    const canonicalBySource = optimizeDataFlow(flow);
-    syncFlowHandles(root, canonicalBySource);
-    flow.initialize();
-
-    /** @type {Promise<void>[]} */
-    const promises = flow.dataSources.map((dataSource) => dataSource.load());
-
-    await Promise.all(promises);
-
-    return flow;
-}
-
-/**
- * Synchronize flow handles after data flow optimization.
- *
- * @param {View} root
- * @param {Map<import("../data/sources/dataSource.js").default, import("../data/sources/dataSource.js").default>} canonicalBySource
- */
-export function syncFlowHandles(root, canonicalBySource) {
-    for (const view of root.getDescendants()) {
-        const handle = view.flowHandle;
-        if (!handle) {
-            continue;
-        }
-
-        const dataSource = handle.dataSource;
-        if (dataSource) {
-            handle.dataSource = canonicalBySource.get(dataSource) ?? dataSource;
-        }
-    }
-}
-
-/**
- * Initializes data flow and marks for a subtree without reinitializing the whole view tree.
- *
- * @param {View} root
- * @param {import("../data/dataFlow.js").default<View>} flow
- * @returns {{
- *     dataFlow: import("../data/dataFlow.js").default<View>,
- *     unitViews: UnitView[],
- *     dataSources: Set<import("../data/sources/dataSource.js").default>,
- *     graphicsPromises: Promise<import("../marks/mark.js").default>[]
- * }}
- */
-export function initializeSubtree(root, flow) {
-    const dataFlow = buildDataFlow(root, flow);
-    const canonicalBySource = optimizeDataFlow(dataFlow);
-    syncFlowHandles(root, canonicalBySource);
-    const subtreeViews = root.getDescendants();
-    /** @type {Set<import("../data/sources/dataSource.js").default>} */
-    const dataSources = new Set();
-    for (const view of subtreeViews) {
-        let current = view;
-        while (current && !current.flowHandle?.dataSource) {
-            current = current.dataParent;
-        }
-        if (current?.flowHandle?.dataSource) {
-            dataSources.add(current.flowHandle.dataSource);
-        }
-    }
-
-    for (const dataSource of dataSources) {
-        dataSource.visit((node) => node.initialize());
-    }
-
-    /** @type {UnitView[]} */
-    const unitViews = subtreeViews.filter((view) => view instanceof UnitView);
-
-    /** @type {Promise<import("../marks/mark.js").default>[]} */
-    const graphicsPromises = [];
-
-    const canInitializeGraphics = !!root.context.glHelper;
-
-    for (const view of unitViews) {
-        const mark = view.mark;
-        mark.initializeEncoders();
-        if (canInitializeGraphics) {
-            graphicsPromises.push(mark.initializeGraphics().then(() => mark));
-        }
-
-        flow.addObserver(
-            view.flowHandle.collector,
-            (collector) => {
-                mark.initializeData(); // does faceting
-                mark.updateGraphicsData();
-            },
-            view.flowHandle
-        );
-    }
-
-    return {
-        dataFlow,
-        unitViews,
-        dataSources,
-        graphicsPromises,
-    };
 }
 
 /**
